+++ vue-components/src/main.ts
@@ -1,4 +1,5 @@
 import Button from './components/Button.vue';
+import Checkbox from './components/Checkbox.vue';
 import TextInput from './components/TextInput.vue';
 import Lookup from './components/Lookup.vue';
 import Dropdown from './components/Dropdown.vue';
@@ -7,6 +8,7 @@ import Icon from './components/Icon.vue';
 
 export {
 	Button,
+	Checkbox,
 	Dropdown,
 	TextInput,
 	Lookup,